requirements.txt
# Add your dependencies here

pymongo
python-dotenv
flask
boto3